current 2025 commit april 6/4/2025

+++ shopitems/src/main/java/org/example/services/payloads/responses/dtos/ItemDTO.java
@@ -14,7 +14,7 @@ public class ItemDTO {
     public BigDecimal costPrice;
     public BigDecimal sellingPrice;
     public String unitOfMeasure;
-    public Integer stockAtHand;
+    public BigDecimal stockAtHand;
 
     public ItemDTO(Item item) {
         this.id = item.id;
